Properly resolved merge conflicts.
